requirements.txt
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-after-response==0.2.2
django-bootstrap4==0.0.7
django-filter==2.1.0
djangorestframework==3.9.2
emoji-data-python==1.1.0
idna==2.8
markdown2==2.3.7
Markdown==3.0.1
psycopg2-binary==2.8.2
python-slugify==1.2.6
pytz==2018.9
requests==2.21.0
six==1.12.0
slackclient==1.3.1
urllib3==1.24.1
websocket-client==0.54.0
